shouts/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.4
django-extensions==3.1.0
djangorestframework==3.12.2
filelock==3.0.12
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
virtualenv==20.2.2
pillow==8.0.1
django-cors-headers==3.6.0
